Rename component to EmailUpdatesModeSelectorComponent
