+++ app/components/op_primer/email_updates_mode_selector_component.html.erb
@@ -28,17 +28,17 @@ See COPYRIGHT and LICENSE files for more details.
 ++#%>
 
 <%=
-  component_wrapper(data: { "test-selector": "notifications-button" }) do
+  component_wrapper(data: { "test-selector": "email-updates-mode-selector" }) do
     render(Primer::OpenProject::SidePanel::Section.new) do |section|
-      section.with_title { I18n.t("meeting.notifications.sidepanel.title") }
+      section.with_title { @title }
 
-      if show_button?
+      if @show_button
         section.with_footer_button(
           color: :accent,
           tag: :a,
-          href: toggle_notifications_dialog_project_meeting_path(@project, @meeting),
+          href: @path,
           classes: "hide-when-print",
-          data: { "turbo-stream": true }
+          data: { "turbo-stream": true, "turbo-method": @method, "test-selector": "email-updates-mode-selector-button" }
         ) do |button|
           button.with_leading_visual_icon(icon: button_icon)
           button_label
@@ -47,13 +47,13 @@ See COPYRIGHT and LICENSE files for more details.
 
       flex_layout do |flex|
         flex.with_row do
-          concat(render(Primer::Beta::Text.new(font_weight: :semibold, color: :subtle)) { "#{state} " })
+          concat(render(Primer::Beta::Text.new(color: :subtle, font_weight: :semibold)) { "#{state} " })
           concat(render(Primer::Beta::Text.new(color: :subtle)) { description })
         end
 
-        unless show_button?
+        unless @show_button
           flex.with_row(mt: 2) do
-            render(Primer::Beta::Text.new(color: :subtle)) { additional_text }
+            render(Primer::Beta::Text.new(color: :subtle)) { @alt_text }
           end
         end
       end
