+++ modules/storages/app/components/storages/admin/side_panel_component.html.erb
@@ -6,7 +6,7 @@
       if @storage.automatic_management_enabled?
         panel.with_section(Storages::Admin::SidePanel::HealthNotificationsComponent.new(storage: @storage))
 
-        panel.with_section(Storages::Admin::SidePanel::EmailUpdatesSwitchComponent.new(storage: @storage))
+        panel.with_section(Storages::Admin::SidePanel::EmailUpdatesModeSelectorComponent.new(storage: @storage))
       end
     end
   end
