+++ modules/storages/app/controllers/storages/admin/storages_controller.rb
@@ -160,7 +160,7 @@ class Storages::Admin::StoragesController < ApplicationController
 
   def change_health_notifications_enabled
     if @storage.update(health_notifications_enabled: !@storage.health_notifications_enabled)
-      update_via_turbo_stream(component: Storages::Admin::SidePanel::EmailUpdatesSwitchComponent.new(storage: @storage))
+      update_via_turbo_stream(component: Storages::Admin::SidePanel::EmailUpdatesModeSelectorComponent.new(storage: @storage))
       respond_with_turbo_streams
     else
       flash.now[:error] = I18n.t("storages.health_email_notifications.error_could_not_be_saved")
